fixes #6757 - composite view to allow duplicate repos
This commit removes the restriction that component content views
added to a composite content view are not allowed to contain
the same source repository.  This is to handle the case where
the user is managing multiple instances of the repos with different
views/purposes and want to combine them through composites.

With this, when publishing a composite content view, if there are
N instances of the same repository, the version published for
the composite will only have 1 repository that combines them
all.

This change in behavior also affects incremental update as it
needs to ensure that that single repository is still a combination
of all versions on the composite (e.g. from new incremental versions
as well as the prior published versions).